+++ lib/hexapdf/font/encoding/base.rb
@@ -61,8 +61,8 @@ module HexaPDF
           @code_to_name.fetch(code, :'.notdef')
         end
 
-        # Returns the Unicode value in UTF-8 for the given code, or an empty string if the code
-        # cannot be mapped.
+        # Returns the Unicode value in UTF-8 for the given code, or +nil+ if the code cannot be
+        # mapped.
         #
         # Note that this method caches the result of the Unicode mapping and therefore should only
         # be called after all codes have been defined.